+++ presentation/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("com.android.application")
+    id("com.android.library")
     id("kotlin-android")
     id("dagger.hilt.android.plugin")
 
@@ -8,20 +8,13 @@ plugins {
 }
 
 android {
-    namespace = "com.beeeam.gocafein"
+    namespace = "com.beeeam.presentation"
     compileSdk = Configuration.COMPILE_SDK
 
     defaultConfig {
-//        applicationId = "com.beeeam.gocafein"
         minSdk = Configuration.MIN_SDK
-        targetSdk = Configuration.TARGET_SDK
-//        versionCode = Configuration.VERSION_CODE
-//        versionName = Configuration.VERSION_NAME
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
     }
 
     buildTypes {
@@ -40,25 +33,47 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.2.0"
+    }
 }
 
 dependencies {
-    implementation(project(":data"))
     implementation(project(":domain"))
-    implementation(project(":presentation"))
 
+    implementation(AndroidX.APP_COMPAT)
     implementation(AndroidX.CORE_KTX)
     implementation(AndroidX.LIFECYCLE_RUNTIME)
+
+    implementation(AndroidX.ACTIVITY_COMPOSE)
+    implementation(AndroidX.COMPOSE_BOM)
+    implementation(AndroidX.COMPOSE_UI)
+    implementation(AndroidX.COMPOSE_UI_GRAPHICS)
+    implementation(AndroidX.COMPOSE_PREVIEW)
+    implementation(AndroidX.COMPOSE_MATERIAL)
+    implementation(AndroidX.NAVIGATION_COMPOSE)
+    implementation(AndroidX.HILT_NAVIGATION)
+
     testImplementation(AndroidX.JUNIT)
     androidTestImplementation(AndroidX.EXT_JUNIT)
     androidTestImplementation(AndroidX.ESPRESSO_CORE)
 
+    androidTestImplementation(platform(AndroidX.COMPOSE_BOM))
+    androidTestImplementation(AndroidX.COMPOSE_UI_JUNIT)
+    debugImplementation(AndroidX.COMPOSE_UI_TOOLING)
+    debugImplementation(AndroidX.COMPOSE_UI_TEST_MANIFEST)
+
     implementation(Google.HILT_ANDROID)
     kapt(Google.HILT_ANDROID_COMPILER)
 
-    implementation(SquareUp.RETROFIT2)
-    implementation(SquareUp.RETROFIT2_CONVERTER_GSON)
-    implementation(SquareUp.OKHTTP3)
-    implementation(SquareUp.OKHTTP3_LOGGING)
-    implementation(SquareUp.OKHTTP3_BOM)
+    implementation(Google.MATERIAL)
+    implementation(Glide.GLIDE)
+
+    implementation(KotlinX.KOTLINX_COROUTINE)
+
+    implementation(Orbit.ORBIT_VIEWMODEL)
+    implementation(Orbit.ORBIT_COMPOSE)
 }
