--- presentation/detail/src/main/java/com/beeeam/detail/DetailContract.kt
@@ -1,7 +1,0 @@
-package com.beeeam.detail
-
-data class DetailState(
-    val sampleState: String = "",
-)
-
-sealed interface DetailSideEffect {}
