--- presentation/navigator/src/main/java/com/beeeam/navigator/MainContract.kt
@@ -1,7 +1,0 @@
-package com.beeeam.navigator
-
-data class MainState(
-    val sampleState: String = "",
-)
-
-sealed interface MainSideEffect {}
