--- presentation/navigator/src/main/java/com/beeeam/navigator/MainNavigator.kt
@@ -1,6 +1,0 @@
-package com.beeeam.navigator
-
-sealed class MainNavigator(val route: String) {
-    object Search : MainNavigator("search")
-    object Detail : MainNavigator("detail")
-}
