--- presentation/navigator/src/main/java/com/beeeam/navigator/MainViewModel.kt
@@ -1,17 +1,0 @@
-package com.beeeam.navigator
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import org.orbitmvi.orbit.Container
-import org.orbitmvi.orbit.ContainerHost
-import org.orbitmvi.orbit.annotation.OrbitExperimental
-import org.orbitmvi.orbit.syntax.simple.blockingIntent
-import org.orbitmvi.orbit.syntax.simple.reduce
-import org.orbitmvi.orbit.viewmodel.container
-import javax.inject.Inject
-
-@HiltViewModel
-class MainViewModel @Inject constructor() : ContainerHost<MainState, MainSideEffect>, ViewModel() {
-    override val container: Container<MainState, MainSideEffect> = container(MainState())
-
-}
