--- presentation/search/src/main/java/com/beeeam/search/SearchContract.kt
@@ -1,10 +1,0 @@
-package com.beeeam.search
-
-import com.beeeam.domain.model.MovieInfo
-
-data class SearchState(
-    val searchValue: String = "",
-    val movieList: List<MovieInfo> = listOf()
-)
-
-sealed interface SearchSideEffect {}
