--- presentation/search/src/main/java/com/beeeam/search/component/SearchBar.kt
@@ -1,45 +1,0 @@
-package com.beeeam.search.component
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.platform.SoftwareKeyboardController
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun GocafeinSearchBar(
-    value: String,
-    keyboardController: SoftwareKeyboardController? = LocalSoftwareKeyboardController.current,
-    onValueChanged: (String) -> Unit,
-    onEnterClicked: (String) -> Unit,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(10.dp)),
-    ) {
-        TextField(
-            value = value,
-            onValueChange = onValueChanged,
-            modifier = Modifier.fillMaxWidth(),
-            singleLine = true,
-            keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Search),
-            keyboardActions = KeyboardActions(
-                onSearch = {
-                    if (!value.isNullOrEmpty()) {
-                        onEnterClicked(value)
-                        keyboardController?.hide()
-                    }
-                }
-            )
-        )
-    }
-}
