+++ presentation/src/main/java/com/beeeam/presentation/navigator/MainActivity.kt
@@ -1,0 +1,38 @@
+package com.beeeam.presentation.navigator
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.beeeam.presentation.navigator.ui.theme.GocafeinTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            GocafeinTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    MainScreen()
+                }
+            }
+        }
+    }
+}
+
+@Preview(apiLevel = 33, showBackground = false)
+@Composable
+fun GreetingPreview() {
+    GocafeinTheme {
+        MainScreen()
+    }
+}
