+++ presentation/src/main/java/com/beeeam/presentation/navigator/MainContract.kt
@@ -1,0 +1,7 @@
+package com.beeeam.presentation.navigator
+
+data class MainState(
+    val sampleState: String = "",
+)
+
+sealed interface MainSideEffect {}
