+++ presentation/src/main/java/com/beeeam/presentation/navigator/MainNavigator.kt
@@ -1,0 +1,6 @@
+package com.beeeam.presentation.navigator
+
+sealed class MainNavigator(val route: String) {
+    object Search : MainNavigator("search")
+    object Detail : MainNavigator("detail")
+}
