+++ presentation/src/main/java/com/beeeam/presentation/navigator/MainViewModel.kt
@@ -1,0 +1,14 @@
+package com.beeeam.presentation.navigator
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import org.orbitmvi.orbit.Container
+import org.orbitmvi.orbit.ContainerHost
+import org.orbitmvi.orbit.viewmodel.container
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor() : ContainerHost<MainState, MainSideEffect>, ViewModel() {
+    override val container: Container<MainState, MainSideEffect> = container(MainState())
+
+}
