+++ presentation/src/main/res/drawable/sample_image.xml
