+++ settings.gradle.kts
@@ -16,8 +16,5 @@ dependencyResolutionManagement {
 rootProject.name = "gocafein"
 include(":data")
 include(":domain")
-include(":presentation")
-include(":presentation:search")
-include(":presentation:detail")
-include(":presentation:navigator")
 include(":app")
+include(":presentation")
